Move .locate() from System to LoaderHooks. This function is called by the Loader before fetch(); it is essential private API between Loader and LoaderHooks.
