+++ src/runtime/modules.js
@@ -119,26 +119,6 @@
       return canonicalizeUrl(name);
     },
 
-    // Should only be called just before 'fetch'
-    locate(load) {
-      load.url = this.locate_(load);
-      return load.url;
-    },
-
-    locate_(load) {
-      var normalizedModuleName = load.name;
-      var asJS = normalizedModuleName + '.js';
-      // Tolerate .js endings
-      if (/\.js$/.test(normalizedModuleName))
-        asJS = normalizedModuleName;
-      if (isAbsolute(asJS))
-        return asJS;
-      var baseURL = load.metadata && load.metadata.baseURL;
-      if (baseURL)
-        return resolveUrl(baseURL, asJS);
-      return asJS;
-    },
-
     get(name) {
       var m = getUncoatedModuleInstantiator(name);
       if (!m)
